fully functioning carousel with animation
